+++ demo/pages/elements/form/FormDemo.ts
@@ -176,7 +176,7 @@ export class FormDemoComponent {
             key: 'entityMultiPicker',
             label: 'Entities',
             required: true,
-            readOnly: true,
+            readOnly: false,
             multiple: true,
             config: {
                 resultsTemplate: EntityPickerResults,